pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Masayuki Sunahara <tamanishi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
PyChromecast = "^9.2.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
